refactor: extract CardExpiringInfo to card_expiring.go, DRY up JSON helpers
- Move CardExpiringInfo type and its methods (DisplayBrand, CardIdentifier,
  extractCardExpiringInfo, formatCardExpiry) into notify/card_expiring.go
  since they're shared across email, SMS, and push channels
- Replace three copy-pasted extractActionType/extractDescription/extractRiskLevel
  functions with a single extractJSONString(raw, key) helper
- email_template.go: 384 → 314 lines
