+++ notify/email_template.go
@@ -184,47 +184,6 @@ func buildPaymentFailedHTMLBody(approval Approval) string {
 	return b.String()
 }
 
-// CardExpiringInfo extracts card details from the Approval.Context JSON
-// for card-expiring notifications.
-type CardExpiringInfo struct {
-	Brand    string `json:"brand"`
-	Last4    string `json:"last4"`
-	Label    string `json:"label,omitempty"` // user-assigned label (e.g. "Work Card")
-	ExpMonth int    `json:"exp_month"`
-	ExpYear  int    `json:"exp_year"`
-	Expired  bool   `json:"expired"` // true if the card is already expired
-}
-
-// DisplayBrand returns the human-readable brand name for card-expiring templates.
-func (c CardExpiringInfo) DisplayBrand() string {
-	return FormatBrand(c.Brand)
-}
-
-// CardIdentifier returns a concise identifier like "Visa ending in 4242" or
-// "Work Card (Visa ending in 4242)" when a label is set. Used in subject lines
-// and notification bodies so users can identify which card needs attention.
-func (c CardExpiringInfo) CardIdentifier() string {
-	base := fmt.Sprintf("%s ending in %s", c.DisplayBrand(), c.Last4)
-	if c.Label != "" {
-		return fmt.Sprintf("%s (%s)", c.Label, base)
-	}
-	return base
-}
-
-// extractCardExpiringInfo pulls card details from the context JSONB.
-func extractCardExpiringInfo(raw json.RawMessage) CardExpiringInfo {
-	var info CardExpiringInfo
-	if len(raw) > 0 {
-		json.Unmarshal(raw, &info) //nolint:errcheck // best-effort
-	}
-	return info
-}
-
-// formatCardExpiry returns "MM/YY" for display.
-func formatCardExpiry(month, year int) string {
-	return fmt.Sprintf("%02d/%02d", month, year%100)
-}
-
 func buildCardExpiringSubject(approval Approval) string {
 	info := extractCardExpiringInfo(approval.Context)
 	if info.Expired {
@@ -329,8 +288,11 @@ func emailRiskBadge(level string) string {
 	)
 }
 
-// extractActionType pulls "type" from the action JSONB.
-func extractActionType(raw json.RawMessage) string {
+// extractJSONString pulls a string value from a top-level key in a JSON
+// object. Returns "" if the key is missing, not a string, or the JSON is
+// invalid. Used to extract fields like "type", "description", and
+// "risk_level" from action/context JSONB without defining full structs.
+func extractJSONString(raw json.RawMessage, key string) string {
 	if len(raw) == 0 {
 		return ""
 	}
@@ -338,7 +300,7 @@ func extractActionType(raw json.RawMessage) string {
 	if json.Unmarshal(raw, &m) != nil {
 		return ""
 	}
-	if v, ok := m["type"]; ok {
+	if v, ok := m[key]; ok {
 		var s string
 		if json.Unmarshal(v, &s) == nil {
 			return s
@@ -347,38 +309,6 @@ func extractActionType(raw json.RawMessage) string {
 	return ""
 }
 
-// extractDescription pulls "description" from the context JSONB.
-func extractDescription(raw json.RawMessage) string {
-	if len(raw) == 0 {
-		return ""
-	}
-	var m map[string]json.RawMessage
-	if json.Unmarshal(raw, &m) != nil {
-		return ""
-	}
-	if v, ok := m["description"]; ok {
-		var s string
-		if json.Unmarshal(v, &s) == nil {
-			return s
-		}
-	}
-	return ""
-}
-
-// extractRiskLevel pulls "risk_level" from the context JSONB.
-func extractRiskLevel(raw json.RawMessage) string {
-	if len(raw) == 0 {
-		return ""
-	}
-	var m map[string]json.RawMessage
-	if json.Unmarshal(raw, &m) != nil {
-		return ""
-	}
-	if v, ok := m["risk_level"]; ok {
-		var s string
-		if json.Unmarshal(v, &s) == nil {
-			return s
-		}
-	}
-	return ""
-}
+func extractActionType(raw json.RawMessage) string  { return extractJSONString(raw, "type") }
+func extractDescription(raw json.RawMessage) string { return extractJSONString(raw, "description") }
+func extractRiskLevel(raw json.RawMessage) string   { return extractJSONString(raw, "risk_level") }
